# Optimise parsing performance