--- a/klog/app/cli/main/cli.go
+++ b/klog/app/cli/main/cli.go
@@ -10,8 +10,6 @@ import (
 	"github.com/jotaen/klog/klog/app/cli"
 	"github.com/jotaen/klog/klog/app/cli/lib"
 	"github.com/jotaen/klog/klog/parser"
-	"github.com/jotaen/klog/klog/parser/engine"
-	"github.com/jotaen/klog/klog/parser/txt"
 	"github.com/jotaen/klog/klog/service"
 	"github.com/jotaen/klog/klog/service/period"
 	kongcompletion "github.com/jotaen/kong-completion"
@@ -69,7 +67,7 @@ func Run(homeDir string, meta app.Meta, isDebug bool, args []string) (int, error
 	ctx := app.NewContext(
 		homeDir,
 		meta,
-		engine.ParallelBatchParser[txt.Block, parser.ParsedRecord, txt.Error]{NumberOfWorkers: runtime.NumCPU()},
+		parser.NewParallelParser(runtime.NumCPU()),
 		lib.CliSerialiser{},
 		isDebug,
 	)