--- a/klog/parser/engine/parallel.go
+++ b/klog/parser/engine/parallel.go
@@ -1,4 +1,4 @@
-package engine2
+package engine
 
 import (
 	"sync"