--- a/klog/parser/engine/parallel.go
+++ b/klog/parser/engine/parallel.go
@@ -4,8 +4,8 @@ import (
 	"sync"
 )
 
-type ParallelParser[Txt any, Int any, Out any, Err any] struct {
-	Serialparser    SerialParser[Txt, Int, Out, Err]
+type ParallelBatchParser[Txt any, Int any, Out any, Err any] struct {
+	SerialParser    SerialParser[Txt, Int, Out, Err]
 	NumberOfWorkers int
 }
 
@@ -14,17 +14,17 @@ type batchResult[Out any, Err any] struct {
 	errs []Err
 }
 
-func (p ParallelParser[Txt, Int, Out, Err]) Parse(txt Txt) ([]Out, []Err) {
-	ints := p.Serialparser.PreProcess(txt)
+func (p ParallelBatchParser[Txt, Int, Out, Err]) Parse(txt Txt) ([]Out, []Err) {
+	allInts := p.SerialParser.PreProcess(txt)
 	// Batch up and dispatch to workers.
 	wg := &sync.WaitGroup{}
-	batches := batch(ints, p.NumberOfWorkers)
+	batches := splitUp(allInts, p.NumberOfWorkers)
 	wg.Add(len(batches))
 	resultChannel := make(chan batchResult[Out, Err])
 	for _, b := range batches {
 		go func(ints []Int) {
 			defer wg.Done()
-			outs, errs := p.Serialparser.parseAll(ints)
+			outs, errs := p.SerialParser.parseAll(ints)
 			resultChannel <- batchResult[Out, Err]{outs, errs}
 		}(b)
 	}
@@ -37,7 +37,7 @@ func (p ParallelParser[Txt, Int, Out, Err]) Parse(txt Txt) ([]Out, []Err) {
 
 	// Collect results.
 	i := 0
-	outs := make([]Out, len(ints))
+	outs := make([]Out, len(allInts))
 	var allErrs []Err
 	for result := range resultChannel {
 		if len(result.errs) > 0 {
@@ -55,7 +55,7 @@ func (p ParallelParser[Txt, Int, Out, Err]) Parse(txt Txt) ([]Out, []Err) {
 	return outs, nil
 }
 
-func batch[T any](slice []T, batchSize int) [][]T {
+func splitUp[T any](slice []T, batchSize int) [][]T {
 	batches := make([][]T, 0, (len(slice)+batchSize-1)/batchSize)
 	for batchSize < len(slice) {
 		slice, batches = slice[batchSize:], append(batches, slice[0:batchSize:batchSize])