--- a/klog/parser/parser.go
+++ b/klog/parser/parser.go
@@ -6,7 +6,9 @@ package parser
 import (
 	"github.com/jotaen/klog/klog"
 	"github.com/jotaen/klog/klog/parser/engine"
+	"runtime"
 	"strings"
+	"sync"
 )
 
 // ParsedRecord is a record along with some meta information which is
@@ -19,6 +21,19 @@ type ParsedRecord struct {
 
 	// Style contains the original styling preferences.
 	Style *Style
+
+	i int
+}
+
+type In struct {
+	engine.Block
+	i int
+}
+
+type Out struct {
+	pr   *ParsedRecord
+	errs []engine.Error
+	i    int
 }
 
 // Parse parses a text into a list of Record datastructures. On success, it returns
@@ -27,21 +42,60 @@ func Parse(recordsAsText string) ([]ParsedRecord, []engine.Error) {
 	blocks := engine.GroupIntoBlocks(engine.Split(recordsAsText))
 	records := make([]ParsedRecord, len(blocks))
 	var allErrs []engine.Error
-	for i, block := range blocks {
-		record, style, errs := parseRecord(block.SignificantLines())
-		if len(errs) > 0 {
-			allErrs = append(allErrs, errs...)
-			continue
+
+	// Set up channels.
+	inChannel := make(chan *In)
+	outChannel := make(chan *Out)
+
+	// Dispatch work.
+	wg := &sync.WaitGroup{}
+	wg.Add(len(blocks))
+	go func() {
+		for i, block := range blocks {
+			inChannel <- &In{block, i}
 		}
-		if block[0].LineEnding != "" {
-			style.LineEnding.Set(block[0].LineEnding)
-		}
-		records[i] = ParsedRecord{
-			Record: record,
-			Block:  block,
-			Style:  style,
+		close(inChannel)
+	}()
+
+	// Spawn workers.
+	numWorkers := runtime.NumCPU()
+	for i := 0; i < numWorkers; i++ {
+		go func() {
+			for block := range inChannel {
+				record, style, errs := parseRecord(block.SignificantLines())
+				out := Out{i: block.i}
+				if len(errs) > 0 {
+					out.errs = errs
+				} else {
+					if block.Block[0].LineEnding != "" {
+						style.LineEnding.Set(block.Block[0].LineEnding)
+					}
+					out.pr = &ParsedRecord{
+						Record: record,
+						Block:  block.Block,
+						Style:  style,
+						i:      block.i,
+					}
+				}
+				outChannel <- &out
+				wg.Done()
+			}
+		}()
+	}
+
+	go func() {
+		wg.Wait()
+		close(outChannel)
+	}()
+
+	for out := range outChannel {
+		if out.errs == nil {
+			records[out.i] = *out.pr
+		} else {
+			allErrs = append(allErrs, out.errs...)
 		}
 	}
+
 	if len(allErrs) > 0 {
 		return nil, allErrs
 	}