--- a/klog/parser/style_test.go
+++ b/klog/parser/style_test.go
@@ -87,7 +87,7 @@ func TestElectStyleDoesNotOverrideSetPreferences(t *testing.T) {
 }
 
 func parseOrPanic(recordsAsText ...string) []ParsedRecord {
-	rs, err := Parse(strings.Join(recordsAsText, ""))
+	rs, err := NewSerialParser().Parse(strings.Join(recordsAsText, ""))
 	if err != nil {
 		panic("Invalid data")
 	}