This PR cuts the parsing time roughly in half:

- All regexes are cached now (~20% speed-up)
- The memory allocations are optimised, so it avoids dynamic allocations where possible (~5% speed-up)
- Records are parsed in parallel, so it utilises all cores of the machine (~20% speed-up)
- The preprocessing algorithm is simplified, and doesn’t rely on regexes anymore (~5% speed-up)

The improvements are only noticeable on larger data-sets, i.e. > 500 records. For smaller ones, it’s the same as before.

|Number of records|Before|After|
|-|-|-|
|10|15ms|15ms|
|100|20ms|20ms|
|1.000|60ms|35ms|
|10.000|480ms|210ms|
|100.000|4.500ms|2.000ms|

(The numbers depend on machine and test data-set, but the overall trend should be representative.)